+++ app/models/concerns/attachable.rb
@@ -1,0 +1,9 @@
+module Concerns
+  module Attachable
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :attachments, as: :attachable, dependent: :destroy
+    end
+  end
+end