+++ spec/factories/attachments.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :attachment do
+    asset
+  end
+end